Fix dependency issues by extracting bing_images subproject
