+++ bing_images/src/main/resources/reference.conf
@@ -1,0 +1,6 @@
+wsd.bing_images {
+    api_key = "missing"
+    api_key = ${?WSP_BING_API_KEY}
+    image_folder = "imgdata/bing"
+    image_folder = ${?WSP_API_BING_IMAGE_FOLDER}
+}
