+++ bing_images/src/main/scala/de/tudarmstadt/lt/wsd/bing_images/BingPhoto.scala
@@ -1,4 +1,4 @@
-package de.tudarmstadt.lt.wsd.common.model.images
+package de.tudarmstadt.lt.wsd.bing_images
 
 /**
   * Created by fide on 20.06.17.
