+++ common/src/main/resources/reference.conf
@@ -41,20 +41,6 @@ scalikejdbc.global.loggingSQLAndTime.stackTraceDepth=10
 wsd {
     result_location = "data/results"
     model_location = "data/models"
-    common {
-
-      bing {
-          api_key = "missing"
-          api_key = ${?WSP_BING_API_KEY}
-          image_folder = "imgdata/bing"
-          image_folder = ${?WSP_API_BING_IMAGE_FOLDER}
-      }
-
-      flickr {
-         api_key = "missing"
-         api_key = ${?WSP_FLICKR_API_KEY}
-      }
-     }
 }
 
 org.slf4j.simplelogger.defaultlog = "debug"
