+++ common/src/main/scala/de/tudarmstadt/lt/wsd/common/Run.scala
@@ -3,23 +3,17 @@ package de.tudarmstadt.lt.wsd.common
 import java.io.File
 import javax.inject.Inject
 
-import akka.actor.ActorSystem
 import skinny.Skinny
-import akka.stream.ActorMaterializer
 import com.typesafe.config.{ConfigFactory, ConfigRenderOptions}
 import com.typesafe.scalalogging.LazyLogging
 import de.tudarmstadt.lt.wsd.common.eval.{EvaluationStatistics, Evaluator, SenseInventoryMapping}
-import de.tudarmstadt.lt.wsd.common.model.images.BingImageDownloader
+
 import de.tudarmstadt.lt.wsd.common.model.{Sense, SenseVectorModel}
 import de.tudarmstadt.lt.wsd.common.prediction._
 import de.tudarmstadt.lt.wsd.common.utils.{FileUtils, TSVUtils, Utils}
 import org.apache.commons.io.FilenameUtils
-import play.api.libs.ws.ahc.AhcWSClient
 import scalikejdbc._
 import scalikejdbc.config._
-
-import scala.concurrent.{Await, Future}
-import scala.util.{Failure, Success}
 
 /**
   * Created by fide on 07.12.16.
@@ -33,7 +27,7 @@ object Run extends LazyLogging {
   object Command extends Enumeration {
     type Name = Value
     val predict, indexdb, evalbabel, evalstats, inventorystats, filterdataset, mappingstats,
-    exportinventory, exporttwsiinventory, extractinventory, downloadimages, nocmd = Value
+    exportinventory, exporttwsiinventory, extractinventory, nocmd = Value
   }
 
   import Command._
@@ -230,18 +224,6 @@ object Run extends LazyLogging {
           .text(s"Name of the model to use")
       )
 
-    cmd("downloadimages").action((_, c) => c.copy(mode = downloadimages)).
-      text("Download images for cache from Bing").
-      children(
-        opt[String]('f', "frequencyFile").required()
-          .action((x,c) => c.copy(frequencyFile = x))
-          .text("A tab seperated CSV file with Sense IDs and their frequencies as columns."),
-        opt[Int]('l', "limit").required()
-          .action((x,c) => c.copy(downloadLimit = Some(x)))
-          .text("Number of maximal downloads")
-      )
-
-
     checkConfig {
       case x if x.mode == nocmd => failure("No command given.")
       case x if x.mode == predict
@@ -256,52 +238,6 @@ object Run extends LazyLogging {
       case _ => success
     }
   }
-
-  def mainDownloadImages(frequencyFile: String, maxOpt: Option[Int]): Unit = {
-    val max = maxOpt.getOrElse(10)
-
-    implicit val system = ActorSystem()
-    implicit val materializer = ActorMaterializer()
-    implicit val wsClient = AhcWSClient()
-
-    val (_, wordFreqs) = TSVUtils.readWithHeaders(frequencyFile, removeIncomplete = true)
-
-    val mostFreq = wordFreqs
-      .sortBy(-_("freq").toInt)
-      .map(_("word").toLowerCase)
-      .distinct
-
-    println(s"STARTING DOWNLOADS (all senses for maximal $max words)\n")
-
-    val downloader = new BingImageDownloader()
-
-    val senses = mostFreq.toStream.flatMap(word => Sense.findAllByCaseIgnoredWord(word))
-    val processed = senses.map { sense =>
-
-      println(s"Checking image for ${sense.uniqueID}.")
-
-      val check = downloader.readFromDiskOrDownload(sense)
-
-      import scala.concurrent.ExecutionContext.Implicits.global
-      check.onComplete {
-        case Success(url) => println(s"Successfully ensured image for ${sense.uniqueID} exists.")
-        case Failure(exception) => println(s"Error for ${sense.uniqueID}: $exception\n")
-      }
-
-      check
-
-    } take max takeWhile { check =>
-      import scala.concurrent.duration._
-      // https://stackoverflow.com/a/21417522
-      Await.ready(check, 1 minutes).value.get.isSuccess
-    } length
-
-    println(s"$processed senses processed")
-
-    wsClient.close()
-    system.terminate()
-  }
-
   def main(args: Array[String]): Unit = {
 
     parser.parse(args, Params()) match {
@@ -326,7 +262,6 @@ object Run extends LazyLogging {
           case `evalstats` => mainEvalStats(params)
           case `exportinventory` => mainExportDefaultInventory(params)
           case `mappingstats` => mainMappingStats(params)
-          case `downloadimages` => mainDownloadImages(params.frequencyFile, params.downloadLimit)
         }
       case None =>
        // Utils.printConfig(config) // TODO why is printing even for common/run
